HostPolicy.Factory -> HostPolicyFactory
